add remoção segura

+++ TP10_TP11/lib/executarExercicio.sh
@@ -42,7 +42,10 @@ function compilar(){
 }
 
 function apagarArquivosGerados(){
-	imprimir "Apagando arquivos gerados"
+	read -n1 -p "Apagar '$PATH_EXEC' e '$PATH_OUTPUT'? (y) "
+	[[ ${REPLY,,} == "y" ]] || exit
+
+	imprimir "\nApagando arquivos gerados"
 	rm -f "$PATH_EXEC"
 	rm -rf "$PATH_OUTPUT"
 }
